Rely on Ruby's flags for the build script (Alexander Kellett).

+++ ext/extconf.rb
@@ -1,28 +1,26 @@
 require 'mkmf'
+require 'rbconfig'
 
 HERE = File.expand_path(File.dirname(__FILE__))
 BUNDLE = Dir.glob("libmemcached-*.tar.gz").first
 BUNDLE_PATH = BUNDLE.sub(".tar.gz", "")
 
-DARWIN = `uname -sp` == "Darwin i386\n"
-
-# Is there a better way to do this?
-archflags = if ENV['ARCHFLAGS']
-  ENV['ARCHFLAGS']
-elsif Config::CONFIG['host_os'][0,8] == "darwin10"
-  "-arch i386 -arch x86_64"
-elsif Config::CONFIG['host_os'][0,6] == "darwin"
-  "-arch i386 -arch ppc"
+if ENV['DEBUG']
+  puts "Setting debug flags."
+  $CFLAGS << " -O0 -ggdb -DHAVE_DEBUG"
+  $EXTRA_CONF = " --enable-debug"
 end
+
+$CFLAGS = "#{RbConfig::CONFIG['CFLAGS']} #{$CFLAGS}".split(" ").uniq.join(" ").gsub("$(cflags)", "")
+$LDFLAGS = "#{RbConfig::CONFIG['LDFLAGS']} #{$LDFLAGS}".split(" ").uniq.join(" ").gsub("$(ldflags)", "")
+$CXXFLAGS = " -std=gnu++98"
+$CPPFLAGS = $ARCH_FLAG = $DLDFLAGS = ""
 
 if !ENV["EXTERNAL_LIB"]
   $includes = " -I#{HERE}/include"
   $libraries = " -L#{HERE}/lib"
-
   $CFLAGS = "#{$includes} #{$libraries} #{$CFLAGS}"
   $LDFLAGS = "#{$libraries} #{$LDFLAGS}"
-  $CPPFLAGS = $ARCH_FLAG = $DLDFLAGS = ""
-
   $LIBPATH = ["#{HERE}/lib"]
   $DEFLIBPATH = []
 
@@ -38,38 +36,17 @@ if !ENV["EXTERNAL_LIB"]
       puts(cmd = "patch -p1 < libmemcached.patch")
       raise "'#{cmd}' failed" unless system(cmd)
 
-      Dir.chdir(BUNDLE_PATH) do
+      Dir.chdir(BUNDLE_PATH) do        
+        puts(cmd = "env CFLAGS='-fPIC #{$CFLAGS}' LDFLAGS='-fPIC #{$LDFLAGS}' ./configure --prefix=#{HERE} --without-memcached --disable-shared --disable-utils #{$EXTRA_CONF} 2>&1")
         
-        cxxflags = cflags = ldflags = "-fPIC"
-        extraconf = ''
-        
-        # again... is there a better way to do this?
-        if DARWIN
-          cflags = "#{cflags} #{archflags}"
-          cxxflags = "-std=gnu++98 #{cflags}"
-          ldflags = "#{ldflags} #{archflags}"
-          extraconf = '--enable-dtrace --disable-dependency-tracking'
-        end
-                
-        if ENV['DEBUG']
-          puts "Setting debug flags for libmemcached."
-          cflags << " -O0 -ggdb -DHAVE_DEBUG"
-          extraconf << " --enable-debug"
-        else
-          cflags << " -Os"
-        end
-        
-        puts(cmd = "env CFLAGS='#{cflags}' LDFLAGS='#{ldflags}' ./configure --prefix=#{HERE} --without-memcached --disable-shared --disable-utils #{extraconf} 2>&1")
         raise "'#{cmd}' failed" unless system(cmd)
-        puts(cmd = "make CXXFLAGS='#{cxxflags}' || true 2>&1")
+        puts(cmd = "make CXXFLAGS=#{$CXXFLAGS} || true 2>&1")
         raise "'#{cmd}' failed" unless system(cmd)
         puts(cmd = "make install || true 2>&1")
         raise "'#{cmd}' failed" unless system(cmd)
       end
 
-      unless ENV['DEBUG'] or ENV['DEV']
-        system("rm -rf #{BUNDLE_PATH}")
-      end
+      system("rm -rf #{BUNDLE_PATH}") unless ENV['DEBUG'] or ENV['DEV']
     end
   end
   
@@ -79,22 +56,6 @@ if !ENV["EXTERNAL_LIB"]
     system("cp -f libmemcached.la libmemcached_gem.la") 
   end
   $LIBS << " -lmemcached_gem"
-end
-
-$CFLAGS.gsub! /-O\d/, ''
-
-if ENV['DEBUG']
-  puts "Setting debug flags for gem."
-  $CFLAGS << " -O0 -ggdb -DHAVE_DEBUG"
-else
-  $CFLAGS << " -Os"
-end
-
-if DARWIN
-  $CFLAGS.gsub! /-arch \S+/, ''
-  $CFLAGS << " #{archflags}"
-  $LDFLAGS.gsub! /-arch \S+/, ''
-  $LDFLAGS << " #{archflags}"
 end
 
 if ENV['SWIG']
